Feat Contact refactor for interactions

+++ src/main/java/com/yestravel/contactsmanager/repo/ContactRepo.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface ContactRepo extends JpaRepository<Contact, Integer> {
+public interface ContactRepo extends JpaRepository<Contact, Long> {
 
     @Override
     Page<Contact> findAll(Pageable pageable);
